requirements.txt
openai
requests
beautifulsoup4
playwright
slack_sdk
notion-client
python-dotenv
schedule
